Renomeada as classes que estavam com o nome errado, estavam como Costumer e o correto é Customer
