+++ backend/java/codebrothers.services.customer/src/main/java/com/codebrothers/services/customer/controllers/CustomerController.java
@@ -1,4 +1,4 @@
-package com.codebrothers.services.customer.service.controllers;
+package com.codebrothers.services.customer.controllers;
 
 import java.net.URI;
 import java.util.List;
@@ -15,28 +15,28 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
-import com.codebrothers.services.customer.entities.Costumer;
-import com.codebrothers.services.customer.service.CostumerService;
+import com.codebrothers.services.customer.entities.Customer;
+import com.codebrothers.services.customer.service.CustomerService;
 
 @RestController
 @RequestMapping(value = "/costumers")
-public class CostumerController {
+public class CustomerController {
    //  Injeção da classe que faz a parte de regras que forem necessarias
 	@Autowired
-	CostumerService costumerService;
+	CustomerService costumerService;
 
 	@GetMapping()
-	public ResponseEntity<List<Costumer>> findAll() {
+	public ResponseEntity<List<Customer>> findAll() {
 		return ResponseEntity.ok().body(costumerService.findAll());
 	}
 
 	@GetMapping(value = "/{id}")
-	public ResponseEntity<Costumer> findById(@PathVariable Long id) {
+	public ResponseEntity<Customer> findById(@PathVariable Long id) {
 		return ResponseEntity.ok().body(costumerService.findById(id));
 	}
 	
 	@PostMapping
-	public ResponseEntity<Costumer> insert(@RequestBody Costumer costumer)
+	public ResponseEntity<Customer> insert(@RequestBody Customer costumer)
 	{
 		costumer = costumerService.insert(costumer);
 		// adiciona ao header uma url com ID do registro criado
@@ -53,7 +53,7 @@ public class CostumerController {
 	}
 	
     @PutMapping(value = "/{id}")
-	public ResponseEntity<Costumer> update(@PathVariable Long id,  @RequestBody Costumer costumer)
+	public ResponseEntity<Customer> update(@PathVariable Long id,  @RequestBody Customer costumer)
 	{
 		costumer = costumerService.update(id, costumer);
 		return ResponseEntity.ok().body(costumer);
