+++ backend/java/codebrothers.services.customer/src/main/java/com/codebrothers/services/customer/exceptions/DataBaseException.java
@@ -1,4 +1,4 @@
-package com.codebrothers.services.customer.service.exceptions;
+package com.codebrothers.services.customer.exceptions;
 /*
  * Classe de exception customizada para o banco
  */
