+++ backend/java/codebrothers.services.customer/src/main/resources/application-dev.properties
@@ -1,10 +1,15 @@
 spring.datasource.url=jdbc:postgresql://localhost:5432/CodeBrothers_Customer
 spring.datasource.username=postgres
 spring.datasource.password=CodeBrother@2020
+# Evitar "warning" do hibernate ao tentar recuperar metadados do postgresql 
 spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation=true
+# Faz atualizações de estrutura de tabelas existentes e cria novas tabelas
 spring.jpa.hibernate.ddl-auto=update
+# Exibe os comandos SQL que forem executados nos logs do console 
 spring.jpa.show-sql=true
+# Exibe os comandos SQL formatados
 spring.jpa.properties.hibernate.format_sql=true
 jwt.secret=MYJWTSECRET
 jwt.expiration=3600000
+# Porta da aplicação
 server.port=8081
